refactor(querybot)🔧!: Refactor QueryBot to remove SQLiteVecDocStore support
- Removed SQLiteVecDocStore import and support from QueryBot.
- Deleted test cases related to SQLiteVecDocStore in QueryBot tests.
- Moved sqlitevec_example.py from examples to archive.
