+++ archive/sqlitevec_example.py
